% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbquartR-package.R
\docType{package}
\name{mbquartR-package}
\alias{mbquartR-package}
\alias{mbquartR}
\title{Finding Manitoba, Canada, quarter sections made easier}
\description{
\code{mbquartR} is an R package for locating quarter sections in Manitoba
using lat/long coordinates or the legal land descriptions using the
Manitoba Original Survey Legal Descriptions data set from the DataMB website
(\url{https://geoportal.gov.mb.ca/}).
}
\details{
There are four main functions in this package:
\enumerate{
\item Download the Manitoba Original Survey Legal Descriptions data set
}
\itemize{
\item \code{\link[=quarters_dl]{quarters_dl()}}
}
\enumerate{
\item Search for legal land descriptions using lat and long coordinates
}
\itemize{
\item \code{\link[=search_coord]{search_coord()}}
}
\enumerate{
\item Search for lat and long coordinates using legal land descriptions
}
\itemize{
\item \code{\link[=search_legal]{search_legal()}}
}
\enumerate{
\item Plot search results on an interactive map
}
\itemize{
\item \code{\link[=map_quarter]{map_quarter()}}
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://docs.ropensci.org/mbquartR/}
  \item \url{https://github.com/ropensci/mbquartR}
  \item Report bugs at \url{https://github.com/ropensci/mbquartR/issues}
}

}
\author{
\strong{Maintainer}: Alex Koiter \email{koitera@brandonu.ca} (\href{https://orcid.org/0000-0002-9355-9561}{ORCID}) (Associate Professor, Department of Geography and Environment, Brandon University <https://alexkoiter.ca/>)

Other contributors:
\itemize{
  \item Emily H Markowitz (Research Fisheries Biologist, Alaska Fisheries Science Center NOAA Fisheries | U.S. Department of Commerce. <https://emilyhmarkowitz.github.io/emilyhmarkowitz> Emily reviewed the package for rOpenSci, see <https://github.com/ropensci/software-review/issues/658>) [reviewer]
  \item Sheila M Saia (Environmental Data Scientist, Tetra Tech. <https://sheilasaia.rbind.io/> Sheila reviewed the package for rOpenSci, see <https://github.com/ropensci/software-review/issues/658>) [reviewer]
}

}
